# Faster prompt eval w/ early exit after last layer's kv cache write