@karpathy 2023-08-07
not willing to bloat things just yet for this speedup