@ochafik 2023-08-07
@karpathy thanks for taking a look (and for the cool repo!). Note that after https://github.com/karpathy/llama2.c/pull/251 it would be a 11-16% speedup for 10 extra LOC (updated figures), but fair enough ✌️